+++ modal-components/src/common/assets/CloseIconSVG.tsx
